+++ demo/demo.clj
@@ -1,13 +1,14 @@
 (ns factual.demo
   (:require [factual.api :as fact]
-            [clojure.data.json :as json]))
+            [clojure.data.json :as json]
+            [sosueme.conf :as conf]))
 
 (defn connect
   "Connects this demo namespace to Factual's API. You must put
    your key and secret in resources/oauth.json.
    See resources/oauth.sample.json for the expected format."
   []
-  (let [{:keys [key secret]} (json/read-json (slurp "resources/oauth.json"))]
+  (let [{:keys [key secret]} (conf/dot-factual "factual-auth.yaml")]
     (fact/factual! key secret)))
 
 (defn simple-fetch